refactor(config): extract to module
