+++ src/cli/config_migrate_glyphs.go
@@ -3,7 +3,7 @@ package cli
 import (
 	"fmt"
 
-	"github.com/jandedobbeleer/oh-my-posh/src/engine"
+	"github.com/jandedobbeleer/oh-my-posh/src/config"
 	"github.com/jandedobbeleer/oh-my-posh/src/platform"
 
 	"github.com/spf13/cobra"
@@ -36,13 +36,13 @@ A backup of the current config can be found at ~/myconfig.omp.json.bak.`,
 	Run: func(_ *cobra.Command, _ []string) {
 		env := &platform.Shell{
 			CmdFlags: &platform.Flags{
-				Config: config,
+				Config: configFlag,
 			},
 		}
 
 		env.Init()
 		defer env.Close()
-		cfg := engine.LoadConfig(env)
+		cfg := config.Load(env)
 
 		cfg.MigrateGlyphs = true
 		if len(format) == 0 {
