+++ src/config/config_test.go
@@ -1,4 +1,4 @@
-package engine
+package config
 
 import (
 	"testing"
@@ -10,25 +10,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	mock2 "github.com/stretchr/testify/mock"
 )
-
-func TestEscapeGlyphs(t *testing.T) {
-	cases := []struct {
-		Input    string
-		Expected string
-	}{
-		{Input: "󰉋", Expected: "\\udb80\\ude4b"},
-		{Input: "a", Expected: "a"},
-		{Input: "\ue0b4", Expected: "\\ue0b4"},
-		{Input: "\ufd03", Expected: "\\ufd03"},
-		{Input: "}", Expected: "}"},
-		{Input: "🏚", Expected: "🏚"},
-		{Input: "\U000F011B", Expected: "\\udb80\\udd1b"},
-		{Input: "󰄛", Expected: "\\udb80\\udd1b"},
-	}
-	for _, tc := range cases {
-		assert.Equal(t, tc.Expected, escapeGlyphs(tc.Input, false), tc.Input)
-	}
-}
 
 func TestGetPalette(t *testing.T) {
 	palette := terminal.Palette{
