+++ src/config/segment_types.go
@@ -1,59 +1,15 @@
-package engine
+package config
 
 import (
 	"errors"
-	"fmt"
-	"runtime/debug"
-	"strings"
-	"time"
 
 	"github.com/jandedobbeleer/oh-my-posh/src/platform"
 	"github.com/jandedobbeleer/oh-my-posh/src/properties"
 	"github.com/jandedobbeleer/oh-my-posh/src/segments"
-	"github.com/jandedobbeleer/oh-my-posh/src/template"
-	"github.com/jandedobbeleer/oh-my-posh/src/terminal"
-
-	c "golang.org/x/text/cases"
-	"golang.org/x/text/language"
 )
 
-// Segment represent a single segment and it's configuration
-type Segment struct {
-	Type                   SegmentType    `json:"type,omitempty" toml:"type,omitempty"`
-	Tips                   []string       `json:"tips,omitempty" toml:"tips,omitempty"`
-	Style                  SegmentStyle   `json:"style,omitempty" toml:"style,omitempty"`
-	PowerlineSymbol        string         `json:"powerline_symbol,omitempty" toml:"powerline_symbol,omitempty"`
-	LeadingPowerlineSymbol string         `json:"leading_powerline_symbol,omitempty" toml:"leading_powerline_symbol,omitempty"`
-	InvertPowerline        bool           `json:"invert_powerline,omitempty" toml:"invert_powerline,omitempty"`
-	Foreground             string         `json:"foreground,omitempty" toml:"foreground,omitempty"`
-	ForegroundTemplates    template.List  `json:"foreground_templates,omitempty" toml:"foreground_templates,omitempty"`
-	Background             string         `json:"background,omitempty" toml:"background,omitempty"`
-	BackgroundTemplates    template.List  `json:"background_templates,omitempty" toml:"background_templates,omitempty"`
-	LeadingDiamond         string         `json:"leading_diamond,omitempty" toml:"leading_diamond,omitempty"`
-	TrailingDiamond        string         `json:"trailing_diamond,omitempty" toml:"trailing_diamond,omitempty"`
-	Template               string         `json:"template,omitempty" toml:"template,omitempty"`
-	Templates              template.List  `json:"templates,omitempty" toml:"templates,omitempty"`
-	TemplatesLogic         template.Logic `json:"templates_logic,omitempty" toml:"templates_logic,omitempty"`
-	Properties             properties.Map `json:"properties,omitempty" toml:"properties,omitempty"`
-	Interactive            bool           `json:"interactive,omitempty" toml:"interactive,omitempty"`
-	Alias                  string         `json:"alias,omitempty" toml:"alias,omitempty"`
-	MaxWidth               int            `json:"max_width,omitempty" toml:"max_width,omitempty"`
-	MinWidth               int            `json:"min_width,omitempty" toml:"min_width,omitempty"`
-	Filler                 string         `json:"filler,omitempty" toml:"filler,omitempty"`
-
-	Enabled bool `json:"-" toml:"-"`
-
-	colors     *terminal.Colors
-	env        platform.Environment
-	writer     SegmentWriter
-	text       string
-	styleCache SegmentStyle
-	name       string
-
-	// debug info
-	duration   time.Duration
-	nameLength int
-}
+// SegmentType the type of segment, for more information, see the constants
+type SegmentType string
 
 // SegmentWriter is the interface used to define what and if to write to the prompt
 type SegmentWriter interface {
@@ -61,26 +17,6 @@ type SegmentWriter interface {
 	Template() string
 	Init(props properties.Properties, env platform.Environment)
 }
-
-// SegmentStyle the style of segment, for more information, see the constants
-type SegmentStyle string
-
-func (s *SegmentStyle) Resolve(env platform.Environment, context any) SegmentStyle {
-	txtTemplate := &template.Text{
-		Context: context,
-		Env:     env,
-	}
-	txtTemplate.Template = string(*s)
-	value, err := txtTemplate.Render()
-	// default to Plain
-	if err != nil || len(value) == 0 {
-		return Plain
-	}
-	return SegmentStyle(value)
-}
-
-// SegmentType the type of segment, for more information, see the constants
-type SegmentType string
 
 const (
 	// Plain writes it without ornaments
@@ -372,102 +308,7 @@ var Segments = map[SegmentType]func() SegmentWriter{
 	YTM:             func() SegmentWriter { return &segments.Ytm{} },
 }
 
-func (segment *Segment) style() SegmentStyle {
-	if len(segment.styleCache) != 0 {
-		return segment.styleCache
-	}
-
-	segment.styleCache = segment.Style.Resolve(segment.env, segment.writer)
-
-	return segment.styleCache
-}
-
-func (segment *Segment) shouldIncludeFolder() bool {
-	if segment.env == nil {
-		return true
-	}
-
-	cwdIncluded := segment.cwdIncluded()
-	cwdExcluded := segment.cwdExcluded()
-
-	return cwdIncluded && !cwdExcluded
-}
-
-func (segment *Segment) isPowerline() bool {
-	style := segment.style()
-	return style == Powerline || style == Accordion
-}
-
-func (segment *Segment) hasEmptyDiamondAtEnd() bool {
-	if segment.style() != Diamond {
-		return false
-	}
-
-	return len(segment.TrailingDiamond) == 0
-}
-
-func (segment *Segment) cwdIncluded() bool {
-	value, ok := segment.Properties[properties.IncludeFolders]
-	if !ok {
-		// IncludeFolders isn't specified, everything is included
-		return true
-	}
-
-	list := properties.ParseStringArray(value)
-
-	if len(list) == 0 {
-		// IncludeFolders is an empty array, everything is included
-		return true
-	}
-
-	return segment.env.DirMatchesOneOf(segment.env.Pwd(), list)
-}
-
-func (segment *Segment) cwdExcluded() bool {
-	value, ok := segment.Properties[properties.ExcludeFolders]
-	if !ok {
-		value = segment.Properties[properties.IgnoreFolders]
-	}
-
-	list := properties.ParseStringArray(value)
-	return segment.env.DirMatchesOneOf(segment.env.Pwd(), list)
-}
-
-func (segment *Segment) shouldInvokeWithTip(tip string) bool {
-	for _, t := range segment.Tips {
-		if t == tip {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (segment *Segment) foreground() string {
-	if segment.colors == nil {
-		segment.colors = &terminal.Colors{}
-	}
-
-	if len(segment.colors.Foreground) == 0 {
-		segment.colors.Foreground = segment.ForegroundTemplates.FirstMatch(segment.writer, segment.env, segment.Foreground)
-	}
-
-	return segment.colors.Foreground
-}
-
-func (segment *Segment) background() string {
-	if segment.colors == nil {
-		segment.colors = &terminal.Colors{}
-	}
-
-	if len(segment.colors.Background) == 0 {
-		segment.colors.Background = segment.BackgroundTemplates.FirstMatch(segment.writer, segment.env, segment.Background)
-	}
-
-	return segment.colors.Background
-}
-
-func (segment *Segment) mapSegmentWithWriter(env platform.Environment) error {
+func (segment *Segment) MapSegmentWithWriter(env platform.Environment) error {
 	segment.env = env
 
 	if segment.Properties == nil {
@@ -486,108 +327,4 @@ func (segment *Segment) mapSegmentWithWriter(env platform.Environment) error {
 	}
 
 	return errors.New("unable to map writer")
-}
-
-func (segment *Segment) string() string {
-	var templatesResult string
-	if !segment.Templates.Empty() {
-		templatesResult = segment.Templates.Resolve(segment.writer, segment.env, "", segment.TemplatesLogic)
-		if len(segment.Template) == 0 {
-			return templatesResult
-		}
-	}
-
-	if len(segment.Template) == 0 {
-		segment.Template = segment.writer.Template()
-	}
-
-	tmpl := &template.Text{
-		Template:        segment.Template,
-		Context:         segment.writer,
-		Env:             segment.env,
-		TemplatesResult: templatesResult,
-	}
-
-	text, err := tmpl.Render()
-	if err != nil {
-		return err.Error()
-	}
-
-	return text
-}
-
-func (segment *Segment) Name() string {
-	if len(segment.name) != 0 {
-		return segment.name
-	}
-
-	name := segment.Alias
-	if len(name) == 0 {
-		name = c.Title(language.English).String(string(segment.Type))
-	}
-
-	segment.name = name
-	return name
-}
-
-func (segment *Segment) SetEnabled(env platform.Environment) {
-	defer func() {
-		err := recover()
-		if err == nil {
-			return
-		}
-		// display a message explaining omp failed(with the err)
-		message := fmt.Sprintf("\noh-my-posh fatal error rendering %s segment:%s\n\n%s\n", segment.Type, err, debug.Stack())
-		fmt.Println(message)
-		segment.Enabled = true
-	}()
-
-	// segment timings for debug purposes
-	var start time.Time
-	if env.Flags().Debug {
-		start = time.Now()
-		segment.nameLength = len(segment.Name())
-		defer func() {
-			segment.duration = time.Since(start)
-		}()
-	}
-
-	err := segment.mapSegmentWithWriter(env)
-	if err != nil || !segment.shouldIncludeFolder() {
-		return
-	}
-
-	segment.env.DebugF("Segment: %s", segment.Name())
-
-	// validate toggles
-	if toggles, OK := segment.env.Cache().Get(platform.TOGGLECACHE); OK && len(toggles) > 0 {
-		list := strings.Split(toggles, ",")
-		for _, toggle := range list {
-			if SegmentType(toggle) == segment.Type || toggle == segment.Alias {
-				return
-			}
-		}
-	}
-
-	if shouldHideForWidth(segment.env, segment.MinWidth, segment.MaxWidth) {
-		return
-	}
-
-	if segment.writer.Enabled() {
-		segment.Enabled = true
-		env.TemplateCache().AddSegmentData(segment.Name(), segment.writer)
-	}
-}
-
-func (segment *Segment) SetText() {
-	if !segment.Enabled {
-		return
-	}
-
-	segment.text = segment.string()
-	segment.Enabled = len(strings.ReplaceAll(segment.text, " ", "")) > 0
-
-	if !segment.Enabled {
-		segment.env.TemplateCache().RemoveSegmentData(segment.Name())
-	}
 }
