--- src/engine/migrate_glyphs_test.go
@@ -1,16 +1,0 @@
-package engine
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestGetCodePoints(t *testing.T) {
-	codepoints, err := getGlyphCodePoints()
-	if connectionError, ok := err.(*ConnectionError); ok {
-		t.Log(connectionError.Error())
-		return
-	}
-	assert.Equal(t, 1939, len(codepoints))
-}
